api/requirements.txt
flask==2.2.2
google.generativeai==0.5.4
flask_cors==4.0.0